Create NewJunction.h
Upper class that both CrossJunction and TJunction inherit from. Unsure of how to initialise class so that It creates cross junction when no data is provided. Same issue with when user determines "Junc_Type", if cross it should create a crossJunction else it should create T junction but not sure how to define that in constructor.